refactored buildPostCard
